LibWeb: Fix interaction between aspect-ratio and box-sizing: border-box
When an element is set with a fixed aspect ratio it is treated as a
replaced element, but in other browsers this behavior is different when
the box-sizing is border box.

For instance, instead of growing the padding outwards, in the case the
element has these properties, it should grow inwards, like a non
replaced element, as its done in other browsers, this is different from
the usual behavior of replaced elements, like images whereas the padding
grows outwards.